backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
opencv-python-headless
pytesseract
brother_ql
Pillow==9.5.0
SQLAlchemy
pdf2image
